create navigation bar template-component and some basic routing
