+++ src/routes/index.js
@@ -1,0 +1,3 @@
+export { Home } from "./home/home.component";
+
+export { Navigation } from "./navigation/navigation.component";
